requirements.txt
#Not needed or per hand:
#pkg-resources==0.0.0
#tensorflow==1.0.0a0

#----
argparse==1.4.0
cycler==0.10.0
dask==0.13.0
decorator==4.0.11
flake8==3.2.1
hacking==0.13.0
ipdb==0.10.2
ipython-genutils==0.1.0
ipython==5.1.0
matplotlib==2.0.0
mccabe==0.5.3
networkx==1.11
numpy==1.12.0
olefile==0.44
pbr==1.10.0
pep8==1.5.7
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.0.0
prompt-toolkit==1.0.9
protobuf==3.1.0.post1
ptyprocess==0.5.1
pycodestyle==2.2.0
pydocstyle==1.1.1
pyflakes==1.3.0
Pygments==2.1.3
pyparsing==2.1.10
python-dateutil==2.6.0
pytz==2016.10
PyYAML==3.12
scikit-image==0.12.3
scipy==0.18.1
simplegeneric==0.8.1
six==1.10.0
toolz==0.8.2
tqdm==4.11.0
traitlets==4.3.1
wcwidth==0.1.7
wheel==0.29.0
